reafactor : 커맨드 디렉토리 추가
